Fix #33: Move macros to a dedicated object.
