+++ build.sbt
@@ -1,4 +1,5 @@
 // shadow sbt-scalajs' crossProject and CrossType until Scala.js 1.0.0 is released
+import com.typesafe.tools.mima.core._
 import sbtcrossproject.{crossProject, CrossType}
 
 val previousVersion = "1.1.0"
@@ -33,6 +34,12 @@ lazy val `portable-scala-reflect` = crossProject(JSPlatform, JVMPlatform, Native
 
     mimaPreviousArtifacts +=
       organization.value %%% moduleName.value % previousVersion,
+    mimaBinaryIssueFilters ++= List(
+      /* Macros were moved to `internal` package so that `Reflect`'s metadata doesn't get polluted by types from `scala-reflect` */
+      ProblemFilters.exclude[DirectMissingMethodProblem]("org.portablescala.reflect.Reflect.lookupInstantiatableClass_impl"),
+      ProblemFilters.exclude[DirectMissingMethodProblem]("org.portablescala.reflect.Reflect.lookupLoadableModuleClass_impl"),
+      ProblemFilters.exclude[MissingClassProblem]("org.portablescala.reflect.Reflect$MacroCompat$*"),
+    ),
 
     publishMavenStyle := true,
     publishTo := {
